refactor(capi): moving non-standard unetsocket apis to a different file
